feat: rename `it_crawler` to `common` and impl `eco.ssu.ac.kr`
