+++ plugins/common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ssufid_itsites"
+name = "ssufid_common"
 version = "0.1.0"
 description.workspace = true
 edition.workspace = true
