+++ plugins/common/src/common/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod gnuboard;
