+++ plugins/common/src/lib.rs
@@ -1,0 +1,3 @@
+mod common;
+
+pub mod sites;
