+++ plugins/common/src/sites/eco.rs
@@ -1,34 +1,35 @@
-use crate::common::it_crawler::ItCrawler;
 use ssufid::{
     PluginError,
     core::{SsufidPlugin, SsufidPost},
 };
 
-pub struct SecPlugin {
-    crawler: ItCrawler<Self>,
+use crate::common::gnuboard::GnuboardCrawler;
+
+pub struct EcoPlugin {
+    crawler: GnuboardCrawler<Self>,
 }
 
-impl SsufidPlugin for SecPlugin {
-    const IDENTIFIER: &'static str = "sec.ssu.ac.kr";
-    const TITLE: &'static str = "숭실대학교 정보보호학과 공지사항";
-    const DESCRIPTION: &'static str = "숭실대학교 정보보호학과 홈페이지의 공지사항을 제공합니다.";
-    const BASE_URL: &'static str = "https://sec.ssu.ac.kr/bbs/board.php?bo_table=notice";
+impl SsufidPlugin for EcoPlugin {
+    const IDENTIFIER: &'static str = "eco.ssu.ac.kr";
+    const TITLE: &'static str = "숭실대학교 경제학과 공지사항";
+    const DESCRIPTION: &'static str = "숭실대학교 경제학과 홈페이지의 공지사항을 제공합니다.";
+    const BASE_URL: &'static str = "https://eco.ssu.ac.kr/bbs/board.php?bo_table=notice";
 
     async fn crawl(&self, posts_limit: u32) -> Result<Vec<SsufidPost>, PluginError> {
         self.crawler.crawl(posts_limit).await
     }
 }
 
-impl Default for SecPlugin {
+impl Default for EcoPlugin {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl SecPlugin {
+impl EcoPlugin {
     fn new() -> Self {
         Self {
-            crawler: ItCrawler::new(),
+            crawler: GnuboardCrawler::new(),
         }
     }
 }
@@ -40,7 +41,7 @@ mod tests {
     #[tokio::test]
     async fn test_crawl() {
         let posts_limit = 100;
-        let plugin = SecPlugin::new();
+        let plugin = EcoPlugin::new();
         let posts = plugin.crawl(posts_limit).await.unwrap();
         assert!(posts.len() <= posts_limit as usize);
         // println!("{:#?}", posts);
