+++ plugins/common/src/sites/sw/graduate.rs
@@ -1,35 +1,36 @@
-use crate::common::it_crawler::ItCrawler;
 use ssufid::{
     PluginError,
     core::{SsufidPlugin, SsufidPost},
 };
 
-pub struct SwBachelorPlugin {
-    crawler: ItCrawler<Self>,
+use crate::common::gnuboard::GnuboardCrawler;
+
+pub struct SwGraduatePlugin {
+    crawler: GnuboardCrawler<Self>,
 }
 
-impl SsufidPlugin for SwBachelorPlugin {
-    const IDENTIFIER: &'static str = "sw.ssu.ac.kr/bachelor";
-    const TITLE: &'static str = "숭실대학교 소프트웨어학부 학사 공지사항";
+impl SsufidPlugin for SwGraduatePlugin {
+    const IDENTIFIER: &'static str = "sw.ssu.ac.kr/graduate";
+    const TITLE: &'static str = "숭실대학교 소프트웨어 대학원 공지사항";
     const DESCRIPTION: &'static str =
-        "숭실대학교 소프트웨어 홈페이지의 학사 공지사항을 제공합니다.";
-    const BASE_URL: &'static str = "https://sw.ssu.ac.kr/bbs/board.php?bo_table=notice";
+        "숭실대학교 컴퓨터학부 홈페이지의 대학원 공지사항을 제공합니다.";
+    const BASE_URL: &'static str = "https://sw.ssu.ac.kr/bbs/board.php?bo_table=gra_notice";
 
     async fn crawl(&self, posts_limit: u32) -> Result<Vec<SsufidPost>, PluginError> {
         self.crawler.crawl(posts_limit).await
     }
 }
 
-impl Default for SwBachelorPlugin {
+impl Default for SwGraduatePlugin {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl SwBachelorPlugin {
+impl SwGraduatePlugin {
     fn new() -> Self {
         Self {
-            crawler: ItCrawler::new(),
+            crawler: GnuboardCrawler::new(),
         }
     }
 }
@@ -41,7 +42,7 @@ mod tests {
     #[tokio::test]
     async fn test_crawl() {
         let posts_limit = 100;
-        let plugin = SwBachelorPlugin::new();
+        let plugin = SwGraduatePlugin::new();
         let posts = plugin.crawl(posts_limit).await.unwrap();
         assert_eq!(posts.len(), posts_limit as usize);
         // println!("{:#?}", posts);
