--- plugins/it_sites/src/common/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod it_crawler;
