--- plugins/it_sites/src/lib.rs
@@ -1,5 +1,0 @@
-mod common;
-
-pub mod cse;
-pub mod sec;
-pub mod sw;
